+++ packages/backend/src/socket/IOProxy.ts
@@ -25,9 +25,10 @@ import {
   ResponseCreateNetworkPayload,
   ErrorCodes,
   AskForMessagesPayload,
-  FileContent,
   FileMetadata,
-  SetChannelSubscribedPayload
+  SetChannelSubscribedPayload,
+  DownloadStatus,
+  RemoveDownloadStatus
 } from '@quiet/state-manager'
 import { emitError } from './errors'
 
@@ -84,20 +85,32 @@ export default class IOProxy {
     await this.getStorage(peerId).sendMessage(message)
   }
 
-  public uploadFile = async (peerId: string, file: FileContent) => {
-    await this.getStorage(peerId).uploadFile(file)
+  public uploadFile = async (peerId: string, metadata: FileMetadata) => {
+    await this.getStorage(peerId).uploadFile(metadata)
   }
 
-  public uploadedFile = (metadata: FileMetadata) => {
-    this.io.emit(SocketActionTypes.UPLOADED_FILE, metadata)
+  public uploadedFile = (payload: FileMetadata) => {
+    this.io.emit(SocketActionTypes.UPLOADED_FILE, payload)
   }
 
   public downloadFile = async (peerId: string, metadata: FileMetadata) => {
     await this.getStorage(peerId).downloadFile(metadata)
   }
 
-  public downloadedFile = (metadata: FileMetadata) => {
-    this.io.emit(SocketActionTypes.DOWNLOADED_FILE, metadata)
+  public cancelDownload = async (peerId: string, mid: string) => {
+    await this.getStorage(peerId).cancelDownload(mid)
+  }
+
+  public updateDownloadProgress = (payload: DownloadStatus) => {
+    this.io.emit(SocketActionTypes.DOWNLOAD_PROGRESS, payload)
+  }
+
+  public removeDownloadStatus = (payload: RemoveDownloadStatus) => [
+    this.io.emit(SocketActionTypes.REMOVE_DOWNLOAD_STATUS, payload)
+  ]
+
+  public updateMessageMedia = (metadata: FileMetadata) => {
+    this.io.emit(SocketActionTypes.UPDATE_MESSAGE_MEDIA, metadata)
   }
 
   // DMs