+++ packages/desktop/src/renderer/sagas/modals/modals.slice.ts
@@ -22,7 +22,6 @@ export class ModalsInitialState {
   [ModalName.joinCommunityModal] = { open: false };
   [ModalName.createCommunityModal] = { open: false };
   [ModalName.uploadedFileModal] = { open: false };
-  [ModalName.unsupportedFileModal] = { open: false };
   [ModalName.sentryWarningModal] = { open: false };
   [ModalName.loadingPanel] = { open: true } // Loading modal is open by default and closes on websocket connection
 }