+++ packages/e2e-tests/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.4-alpha.1](https://github.com/ZbayApp/monorepo/compare/e2e-tests@0.9.4-alpha.0...e2e-tests@0.9.4-alpha.1) (2022-07-15)
+
+**Note:** Version bump only for package e2e-tests
+
+
+
+
+
+## [0.9.4-alpha.0](https://github.com/ZbayApp/monorepo/compare/e2e-tests@0.9.3-alpha.0...e2e-tests@0.9.4-alpha.0) (2022-07-07)
+
+**Note:** Version bump only for package e2e-tests
+
+
+
+
+
+## [0.9.3-alpha.0](https://github.com/ZbayApp/monorepo/compare/e2e-tests@0.9.2-alpha.0...e2e-tests@0.9.3-alpha.0) (2022-07-06)
+
+**Note:** Version bump only for package e2e-tests
+
+
+
+
+
+## [0.9.2-alpha.0](https://github.com/ZbayApp/monorepo/compare/e2e-tests@0.9.1-alpha.0...e2e-tests@0.9.2-alpha.0) (2022-07-05)
+
+**Note:** Version bump only for package e2e-tests
+
+
+
+
+
+## [0.9.1-alpha.0](https://github.com/ZbayApp/monorepo/compare/e2e-tests@0.9.0-alpha.0...e2e-tests@0.9.1-alpha.0) (2022-07-05)
+
+**Note:** Version bump only for package e2e-tests
+
+
+
+
+
+# [0.9.0-alpha.0](https://github.com/ZbayApp/monorepo/compare/e2e-tests@0.8.2-alpha.0...e2e-tests@0.9.0-alpha.0) (2022-07-05)
+
+**Note:** Version bump only for package e2e-tests
+
+
+
+
+
 ## [0.8.2-alpha.0](https://github.com/ZbayApp/monorepo/compare/e2e-tests@0.8.1-alpha.1...e2e-tests@0.8.2-alpha.0) (2022-06-21)
 
 **Note:** Version bump only for package e2e-tests